+++ src/Handler/LogLevelStrategy/FixedStrategy.php
@@ -17,6 +17,7 @@ final class FixedStrategy implements LogLevelStrategyInterface
     /**
      * The default LogLevel that it will used for the non-exceptions.
      *
+     * @phpstan-var LogLevel::*
      * @var string
      */
     private $defaultLevel;
@@ -24,6 +25,7 @@ final class FixedStrategy implements LogLevelStrategyInterface
     /**
      * The log Level in cases of exceptions.
      *
+     * @phpstan-var LogLevel::*
      * @var string
      */
     private $exceptionLevel;
@@ -31,10 +33,16 @@ final class FixedStrategy implements LogLevelStrategyInterface
     /**
      * The log level of the statistics
      *
+     * @phpstan-var LogLevel::*
      * @var string
      */
     private $statsLevel;
 
+    /**
+     * @phpstan-param LogLevel::* $defaultLevel
+     * @phpstan-param LogLevel::*|null $exceptionLevel
+     * @phpstan-param LogLevel::*|null $statsLevel
+     */
     public function __construct(
         string $defaultLevel = LogLevel::DEBUG,
         string $exceptionLevel = null,
@@ -48,9 +56,7 @@ final class FixedStrategy implements LogLevelStrategyInterface
     /**
      * Returns the log level.
      *
-     * @param MessageInterface|\Exception|TransferStats $value
-     * @param array $options
-     * @return string
+     * @param MessageInterface|\Throwable|TransferStats $value
      */
     public function getLevel($value, array $options): string
     {
@@ -58,7 +64,7 @@ final class FixedStrategy implements LogLevelStrategyInterface
             return $this->defaultLevel;
         }
 
-        if ($value instanceof \Exception) {
+        if ($value instanceof \Throwable) {
             return $this->exceptionLevel;
         }