+++ src/Handler/StringHandler.php
@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 namespace GuzzleLogMiddleware\Handler;
 
-use Exception;
 use GuzzleHttp\TransferStats;
 use GuzzleLogMiddleware\Handler\LogLevelStrategy\LogLevelStrategyInterface;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Log\LoggerInterface;
+use Throwable;
 
 /**
  * @author George Mponos <gmponos@gmail.com>
@@ -24,10 +24,10 @@ final class StringHandler extends AbstractHandler
     public function log(
         LoggerInterface $logger,
         RequestInterface $request,
-        ?ResponseInterface $response,
-        ?Exception $exception,
-        ?TransferStats $stats,
-        array $options
+        ?ResponseInterface $response = null,
+        ?Throwable $exception = null,
+        ?TransferStats $stats = null,
+        array $options = []
     ): void {
         $this->logRequest($logger, $request, $options);
 
@@ -73,7 +73,7 @@ final class StringHandler extends AbstractHandler
         $logger->log($level, 'Guzzle HTTP response:' . "\n" . $str);
     }
 
-    private function logReason(LoggerInterface $logger, Exception $exception, array $options): void
+    private function logReason(LoggerInterface $logger, Throwable $exception, array $options): void
     {
         $level = $this->logLevelStrategy->getLevel($exception, $options);
         $logger->log($level, sprintf('Guzzle HTTP exception: %s', $exception->getMessage()), [